refactor: multi-module 작업 - coordilayout 모듈 추가
