+++ build.gradle
@@ -2,8 +2,8 @@ buildscript {
     apply from: 'version.gradle'
     addRepos(repositories)
     dependencies {
-        classpath deps.androidPlugin
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.7.10"
+        classpath (deps.androidPlugin)
+        classpath (deps.kotlinPlugin)
         classpath "com.github.dcendents:android-maven-gradle-plugin:2.1"
     }
 }
