+++ settings.gradle
@@ -7,3 +7,4 @@ include ':viewpager'
 include ':extensions'
 include ':progress'
 include ':tablayout'
+include ':coordinatorlayout'
