+++ widget/src/main/res/values/attrs.xml
@@ -1,37 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <declare-styleable name="TranslationBehavior">
-        <!-- [s] BasePercentageBehavior Attr -->
-        <attr name="behaviorDependId" format="reference" /> <!-- Dependency View Id -->
-        <attr name="behaviorDependType" format="enum"> <!-- Dependency View Type -->
-            <enum name="horizontal" value="0" />
-            <enum name="vertical" value="1" />
-        </attr>
-        <attr name="behaviorDependPin" format="dimension" /> <!-- Dependency Pin Height or Width-->
-        <attr name="behaviorDependRange" format="dimension" /> <!-- Dependency Action Range -->
-        <!-- [e] BasePercentageBehavior Attr -->
-
-        <!-- [s] TranslationBehavior Attr -->
-        <!-- Child View End X Location
-            Format {$Standard , End Location X}
-            @param Standard -> s or e (없어도 됨.)
-            @param End Location X -> 33(dp) s or e 기준으로 Child View 를 마지막에 위치 하고 싶은 Y 좌표.
-            ex.) 23 or s,33
-         -->
-        <attr name="behaviorEndX" format="string" />
-        <!-- Child View End Y Location
-            Format {$Standard , End Location Y}
-            @param Standard -> s or e (없어도 됨.)
-            @param End Location Y -> 33(dp) s or e 기준으로 Child View 를 마지막에 위치 하고 싶은 Y 좌표.
-            ex.) 23 or s,33
-         -->
-        <attr name="behaviorEndY" format="string" />
-        <attr name="behaviorEndAlpha" format="float" /> <!-- Child View End Alpha -->
-
-        <attr name="behaviorEndWidth" format="dimension" />   <!-- Child View End Width -->
-        <attr name="behaviorEndHeight" format="dimension" /> <!-- Child View End Height -->
-        <attr name="behaviorIsFitsSystemWindow" format="boolean"/> <!-- Parent FitsSystemWindow -->
-        <!-- [e] TranslationBehavior Attr -->
-    </declare-styleable>
 
 </resources>
